Deprecate add_to_class decorator
