+++ src/scm/plams/mol/identify.py
@@ -217,77 +217,6 @@ def molecule_name(molecule):
     return sha256(" ".join(names))
 
 
-@add_to_class(Molecule)
-def label(self, level: int = 1, keep_labels: bool = False, flags: Optional[Dict[str, bool]] = None) -> str:
-    """Compute the label of this molecule using chosen *level* of detail.
-
-    Possible levels are:
-
-    *   **0**: does not perform any atom labeling, returns empirical formula (see :meth:`~scm.plams.mol.molecule.Molecule.get_formula`)
-    *   **1**: only direct connectivity is considered, without bond orders (in other words, treats all the bonds as single bonds)
-    *   **2**: use connectivity and bond orders
-    *   **3**: use connectivity, bond orders and some spatial information to distinguish R/S and E/Z isomers
-    *   **4**: use all above, plus more spatial information to distinguish different rotamers and different types of coordination complexes
-
-    If you need more precise control of what is taken into account while computing the label (or adjust the tolerance for geometrical operations) you can use the *flags* argument. It should be a dictionary of parameters recognized by :func:`~scm.plams.mol.identify.label_atoms`. Each of two letter boolean flags has to be present in *flags*. If you use *flags*, *level* is ignored.
-
-    The *level* argument can also be a tuple of integers. In that case the labeling algorithm is run multiple times and the returned value is a tuple (with the same length as *level*) containing labels calculated with given levels of detail.
-
-    This function, by default, erases ``IDname`` attributes of all atoms at the end. You can change this behavior with *keep_labels* argument.
-
-    If the molecule does not contain bonds, :meth:`~scm.plams.mol.molecule.Molecule.guess_bonds` is used to determine them.
-
-
-    .. note::
-
-        This method is a new PLAMS feature and it's still somewhat experimental. The exact details of the algorithm can, and probably will, change in future. You are more than welcome to provide any feedback or feature requests.
-
-    """
-    if isinstance(level, (tuple, list)):
-        return tuple(self.label(i) for i in level)
-
-    if flags is None:
-        if level == 0:
-            return self.get_formula()
-
-        flags = {i: False for i in possible_flags}
-        if level >= 2:
-            flags["BO"] = True
-        if level >= 3:
-            flags["RS"] = True
-            flags["EZ"] = True
-        if level >= 4:
-            flags["DH"] = True
-            flags["CO"] = True
-
-    if len(self.bonds) == 0:
-        self.guess_bonds()
-
-    clear(self)
-    label_atoms(self, **flags)
-    ret = molecule_name(self)
-    if not keep_labels:
-        clear(self)
-    return ret
-
-
-@add_to_class(Molecule)
-def set_local_labels(self, niter=2, flags=None):
-    """
-    Set atomic labels (IDnames) that are unique for local structures of a molecule
-
-    * ``niter`` -- The number of iterations in the atom labeling scheme
-
-    The idea of this method is that the number of iterations can be specified.
-    If kept low (default niter), local structures over different molecules will have the same label.
-    """
-    if flags is None:
-        flags = {i: False for i in possible_flags}
-    initialize(self)
-    for i in range(niter):
-        iterate(self, flags)
-
-
 @requires_optional_package("networkx")
 def get_graph(mol, dic, level=1):
     """
@@ -319,59 +248,3 @@ def get_graph(mol, dic, level=1):
     graph = networkx.from_numpy_array(matrix)
 
     return graph
-
-
-@add_to_class(Molecule)
-@requires_optional_package("networkx")
-def find_permutation(self, other, level=1):
-    """
-    Reorder atoms in this molecule to match the order in some *other* molecule. The reordering is applied only if the perfect match is found. Returned value is the applied permutation (as a list of integers) or ``None``, if no reordering was performed.
-    """
-    import networkx
-
-    # Get bonds and unique atomIDs if needed
-    if len(self.bonds) == 0:
-        self.guess_bonds()
-    if not hasattr(self.atoms[0], "IDname"):
-        self.label(level=1, keep_labels=True)
-
-    # Link atom IDs to integers
-    dic: Dict[str, int] = {}
-    for at in self.atoms:
-        if not at.IDname in dic.keys():
-            dic[at.IDname] = max([v for v in dic.values()]) + 1 if len(dic) > 0 else 1
-
-    # Create the graphs
-    graph = get_graph(self, dic, level=1)
-    graph2 = get_graph(other, dic, level=1)
-    if graph2 is None:
-        return None
-
-    # Match
-    GM = networkx.isomorphism.GraphMatcher(
-        graph, graph2, edge_match=networkx.isomorphism.categorical_edge_match("weight", 1)
-    )
-    isomorphic = GM.is_isomorphic()
-    if not isomorphic:
-        return None
-
-    # Invert the solution dictionary, to be able to reorder the first graph
-    dic = {}
-    for k, v in GM.mapping.items():
-        dic[v] = k
-    keys = sorted([key for key in dic.keys()])
-    indices = [dic[key] for key in keys]
-
-    return indices
-
-
-@add_to_class(Molecule)
-def reorder(self, other, level=1):
-    """
-    Reorder atoms in this molecule to match the order in some *other* molecule. The reordering is applied only if the perfect match is found. Returned value is a new ``Molecule`` object, or ``None`` if no reordering was performed. See also :func:`~scm.plams.mol.identify.find_permutation`.
-    """
-    indices = self.find_permutation(other, level=level)
-    if indices is None:
-        return None
-    mol = self.get_fragment(indices)
-    return mol
